Change form button type from `submit` to `button`
This instructs the browser to NOT submit the form (and cause the page
to reload) the button is located within. This allows us to remove the
JS that exists solely to prevent the same thing.

I started out converting the form to a div, but then all the styling
broke and when I saw how many css lines have `form` in them, I said
"nevermind..." lol